Code refactoring on renaming module
