+++ app/src/main/java/nitmeghalaya/shishir2020/di/KoinModule.kt
@@ -3,7 +3,7 @@ package nitmeghalaya.shishir2020.di
 import nitmeghalaya.shishir2020.repository.FacebookPageRepository
 import nitmeghalaya.shishir2020.repository.FirestoreRepository
 import nitmeghalaya.shishir2020.screens.eventslist.EventsListViewModel
-import nitmeghalaya.shishir2020.screens.shishirpagefeed.ShishirPageFeedViewModel
+import nitmeghalaya.shishir2020.screens.facebookpagefeed.ShishirPageFeedViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
