+++ app/src/main/java/nitmeghalaya/shishir2020/screens/facebookpagefeed/ShishirPageFeedFragment.kt
@@ -1,4 +1,4 @@
-package nitmeghalaya.shishir2020.screens.shishirpagefeed
+package nitmeghalaya.shishir2020.screens.facebookpagefeed
 
 import android.os.Bundle
 import android.view.LayoutInflater
