+++ app/src/main/res/layout/fragment_shishir_page_feed.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.shishirpagefeed.ShishirPageFeedFragment">
+    tools:context=".screens.facebookpagefeed.ShishirPageFeedFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
