+++ app/src/main/res/navigation/nav_graph.xml
@@ -25,7 +25,7 @@
     </fragment>
     <fragment
         android:id="@+id/shishirPageFeedFragment"
-        android:name="nitmeghalaya.shishir2020.screens.shishirpagefeed.ShishirPageFeedFragment"
+        android:name="nitmeghalaya.shishir2020.screens.facebookpagefeed.ShishirPageFeedFragment"
         android:label="@string/feed"
         tools:layout="@layout/fragment_shishir_page_feed" >
         <action
